// src/app/api/actuator/model/serverRequestExecutionStat.ts
/**
 * View API
 * View API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@it4all.hu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ServerRequestType } from './serverRequestType';
import { StatisticRecord } from './statisticRecord';


/**
 * The server request execution statistic is an object for the actuator infrastructure. It is initiated for every view-widget-action and contains the minimum, average and maximum execution times. It collects from the server start. 
 */
export interface ServerRequestExecutionStat { 
    /**
     * The identifier of the statistical entry. The view - widget - action format.
     */
    id?: string;
    /**
     * The name of the view the request is performed on.
     */
    viewName?: string;
    /**
     * The identifier of the widget if it is a widget action perform.
     */
    widgetId?: string;
    /**
     * The identifier of the action performed.
     */
    actionCode?: string;
    fullStat?: StatisticRecord;
    writeCount?: StatisticRecord;
    writeStat?: StatisticRecord;
    readCount?: StatisticRecord;
    readStat?: StatisticRecord;
    type?: ServerRequestType;
}